GraphicsBufferType: Fix Modifier return type from uint32_t to uint64_t

+++ Source/graphicsbuffer/include/graphicsbuffer/GraphicsBufferType.h
@@ -282,7 +282,7 @@ namespace Graphics {
             {
                 return (_format);
             }
-            uint32_t Modifier() const
+            uint64_t Modifier() const
             {
                 return (_modifier);
             }
